Created API call to create a new organization.
The functionality first creates a group using the same static function call
that the API for creating a group uses.  Then it adds the extra information to
turn it into an organization.

The only extra parameter that is sent to the API call is a string 'groupType'.
You can pass a group type that is not specified in the AppConfig.xml
'enum_typed_group_types' section, but it will not be clickable to see other groups
of that type.  It is probably not visible at all unless viewing all organizations
on one page as we are doing for the CivicCommons project.