Updated regex for url validation and setup redirection for contributions.
TODO: The regex is scattered a few places in the code <= booooooo!

If the user is creating a Contribution and the redirection is set (which it will
be from CC), then I parse the query parameters and add the title and link to the
newly created Contribution.  Then I box up the query parameters onto the previous
redirect url and redirect to it.

All urls that are passed as query variables are urlencoded, and will need to be
decoded before storing them on the core side of CC.

Example:
If you go to the following URL as a logged in PA user (or you use the authToken),
then you will be creating a new Contribution.

http://www.paserver.com/post_content.php?blog_type=Contribution&redirect=http%3A%2F%2Fwww.coreserver.com%2Fcontributions%2Fcreate_from_pa%3Fissue_id%3D1%26person_id%3D1

It will be setup to redirect to:
http://www.coreserver.com/contributions/create_from_pa?issue_id=1&person_id=1

I will add the post title and a link to it as query args, so the final redirect
URL will look something like this:
http://www.coreserver.com/contributions/create_from_pa?issue_id=1&person_id=1&title=My%20Contribution!&link=http%3A%2F%2Fwww.paserver.com%2Fcontent%2Fcid%3D9