Created new paproject files for displaying a group.
Warning:
I just copied the code from the header_user.tpl file so there may be things that
are needed from the old header_group.tpl file that I did not include. For the
Civic Commons install, we do not allow the user to change the group theme or header
image so we probably avoid many of the problems for now but it is something to note.